requirements.txt
python-telegram-bot --pre
python-dotenv
# No uvicorn/fastapi needed as we removed the dashboard
